Simplify topic handler.
1. We don't need to know whether a command was addressed to an individual
panel or a room (or later a group). That means that it is enough to match
on the last part that specifies the channel.

2. We therefore do not need to keep the cmnd/... topic strings around to
check against. We only build them to subscribe to them and then forget
them.

3. cmnd_map and stat_map now always have the entry for command/stat_event_t
no. i at index i. There is no need for a struct, just store the string.